Split test module into target and transformation tests
